+++ .github/workflows/deploy.yml
@@ -29,7 +29,7 @@ jobs:
         workflow_conclusion: completed
         pr: ${{github.event.pull_request.number}}
         name: wasm
-        path: ./public
+        path: ./public/built
 
     - name: Deploy to Github Pages
       uses: peaceiris/actions-gh-pages@v3
