+++ bindings/wasm/examples/built/manifold-encapsulated-types.d.ts
