+++ bindings/wasm/examples/built/manifold.wasm
