+++ bindings/wasm/examples/three.html
@@ -54,7 +54,7 @@
 <script type="module">
   import * as THREE from 'https://cdn.jsdelivr.net/npm/three@0.149.0/build/three.module.js';
   import { mergeVertices } from 'https://cdn.jsdelivr.net/npm/three@0.149.0/examples/jsm/utils/BufferGeometryUtils.js';
-  import Module from './manifold.js';
+  import Module from './built/manifold.js';
 
   const wasm = await Module();
   wasm.setup();
